# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShopFlow API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/products": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "List active products (cursor-paginated)",
        "operationId": "list_products_products_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProductResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "products"
        ],
        "summary": "Create a product (merchant)",
        "operationId": "create_product_products_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a merchant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/orders": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "List orders visible to the caller (cursor-paginated)",
        "operationId": "list_orders_orders_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_OrderResponse_"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/merchant/dashboard": {
      "get": {
        "tags": [
          "merchant"
        ],
        "summary": "Merchant dashboard: revenue windows, orders by status, top products, low stock",
        "operationId": "dashboard_merchant_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantDashboard"
                }
              }
            }
          },
          "403": {
            "description": "Not a merchant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/merchant/products": {
      "get": {
        "tags": [
          "merchant"
        ],
        "summary": "List the merchant's own products, all statuses (cursor-paginated)",
        "operationId": "list_products_merchant_products_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProductStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProductResponse_"
                }
              }
            }
          },
          "403": {
            "description": "Not a merchant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/merchant/orders": {
      "get": {
        "tags": [
          "merchant"
        ],
        "summary": "List orders containing the merchant's products (cursor-paginated)",
        "operationId": "list_orders_merchant_orders_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OrderStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_OrderResponse_"
                }
              }
            }
          },
          "403": {
            "description": "Not a merchant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/merchant/revenue-summary": {
      "get": {
        "tags": [
          "merchant"
        ],
        "summary": "Merchant revenue summary (delivered orders only)",
        "operationId": "revenue_summary_merchant_revenue_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevenueSummary"
                }
              }
            }
          },
          "403": {
            "description": "Not a merchant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LowStockProduct": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "stock_qty": {
            "type": "integer",
            "title": "Stock Qty"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "title",
          "stock_qty"
        ],
        "title": "LowStockProduct"
      },
      "MerchantDashboard": {
        "properties": {
          "revenue": {
            "$ref": "#/components/schemas/RevenueWindows"
          },
          "orders_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Orders By Status"
          },
          "top_products": {
            "items": {
              "$ref": "#/components/schemas/TopProduct"
            },
            "type": "array",
            "title": "Top Products"
          },
          "low_stock": {
            "items": {
              "$ref": "#/components/schemas/LowStockProduct"
            },
            "type": "array",
            "title": "Low Stock"
          },
          "product_count": {
            "type": "integer",
            "title": "Product Count"
          }
        },
        "type": "object",
        "required": [
          "revenue",
          "orders_by_status",
          "top_products",
          "low_stock",
          "product_count"
        ],
        "title": "MerchantDashboard"
      },
      "OrderItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Unit Price"
          },
          "line_total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Line Total"
          }
        },
        "type": "object",
        "required": [
          "id",
          "product_id",
          "quantity",
          "unit_price",
          "line_total"
        ],
        "title": "OrderItemResponse"
      },
      "OrderResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "customer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Customer Id"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "total_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Amount"
          },
          "shipping_address": {
            "$ref": "#/components/schemas/ShippingAddress"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer_id",
          "status",
          "total_amount",
          "shipping_address",
          "created_at",
          "items"
        ],
        "title": "OrderResponse"
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "pending",
          "confirmed",
          "shipped",
          "delivered",
          "cancelled"
        ],
        "title": "OrderStatus"
      },
...
      "Page_OrderResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[OrderResponse]"
      },
      "Page_ProductResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[ProductResponse]"
      },
...
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "A URI reference identifying the problem type.",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "A short, human-readable summary of the problem type."
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "The HTTP status code."
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "A human-readable explanation specific to this occurrence."
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance",
            "description": "A URI reference identifying the specific occurrence."
          }
        },
        "type": "object",
        "required": [
          "title",
          "status"
        ],
        "title": "ProblemDetail",
        "description": "RFC 7807 Problem Details body (media type ``application/problem+json``).\n\nUsed as the documented response model for every error in the OpenAPI spec.",
        "example": {
          "detail": "Invalid or expired credentials.",
          "instance": "/auth/login",
          "status": 401,
          "title": "Unauthorized",
          "type": "about:blank"
        }
      },
      "ProductCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,11}0*$)\\d{0,8}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Price"
          },
          "stock_qty": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Stock Qty",
            "default": 0
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Images"
          },
          "status": {
            "$ref": "#/components/schemas/ProductStatus",
            "default": "draft"
          }
        },
        "type": "object",
        "required": [
          "title",
          "price"
        ],
        "title": "ProductCreate"
      },
      "ProductResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "merchant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Merchant Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price"
          },
          "stock_qty": {
            "type": "integer",
            "title": "Stock Qty"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Images"
          },
          "status": {
            "$ref": "#/components/schemas/ProductStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "merchant_id",
          "title",
          "description",
          "price",
          "stock_qty",
          "category_id",
          "images",
          "status"
        ],
        "title": "ProductResponse"
      },
      "ProductStatus": {
        "type": "string",
        "enum": [
          "draft",
          "active",
          "archived"
        ],
        "title": "ProductStatus"
      },
...
      "RevenueSummary": {
        "properties": {
          "total_revenue": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Revenue"
          },
          "delivered_order_count": {
            "type": "integer",
            "title": "Delivered Order Count"
          },
          "average_order_value": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Average Order Value"
          },
          "revenue": {
            "$ref": "#/components/schemas/RevenueWindows"
          }
        },
        "type": "object",
        "required": [
          "total_revenue",
          "delivered_order_count",
          "average_order_value",
          "revenue"
        ],
        "title": "RevenueSummary"
      },
      "RevenueWindows": {
        "properties": {
          "last_7_days": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Last 7 Days"
          },
          "last_30_days": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Last 30 Days"
          },
          "last_90_days": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Last 90 Days"
          },
          "all_time": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "All Time"
          }
        },
        "type": "object",
        "required": [
          "last_7_days",
          "last_30_days",
          "last_90_days",
          "all_time"
        ],
        "title": "RevenueWindows",
        "description": "Revenue rolled up over trailing time windows plus an all-time total."
      },
...
      "ShippingAddress": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Full Name"
          },
          "line1": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Line1"
          },
          "line2": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Line2"
          },
          "city": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "City"
          },
          "state": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "postal_code": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Postal Code"
          },
          "country": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "Country"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "line1",
          "city",
          "postal_code",
          "country"
        ],
        "title": "ShippingAddress"
      },
      "TopProduct": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "units_sold": {
            "type": "integer",
            "title": "Units Sold"
          },
          "revenue": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Revenue"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "title",
          "units_sold",
          "revenue"
        ],
        "title": "TopProduct"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}